requirements.txt
Flask==2.3.3
yfinance==0.2.18
pandas==2.0.3
numpy==1.24.3
talib-binary
plotly==5.17.0
requests==2.31.0
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2